Just need to work on filter

+++ trash-stack.server/server/services/NotesService.js
@@ -2,8 +2,16 @@ import { dbContext } from '../db/DbContext'
 import { BadRequest } from '../utils/Errors'
 
 class NotesService {
-  async delete(id) {
-    const deletedNote = await dbContext.Note.findOneAndDelete(id)
+  async updateNote(id, body) {
+    const updatedNote = await dbContext.Note.findOneAndUpdate(id, body)
+    if (!updatedNote) {
+      throw new BadRequest('Need a different Id you gave me' + `${id}` + 'or you gave the wrong body info' + `${body}`)
+    }
+    return updatedNote
+  }
+
+  async delete(id, cId) {
+    const deletedNote = await dbContext.Note.findOneAndDelete({ _id: id, creatorId: cId })
     if (!deletedNote) {
       throw new BadRequest('You may have the wrong Id' + `${id}`)
     }
